fixing admin side

+++ application/modules/admin/controllers/Category.php
@@ -157,13 +157,18 @@ public function create_action()
     {
         $row = $this->Category_model->get_by_id($id);
         $check_sub_cat= $this->Category_model->check_sub_category_id($id);
+        $check_product = $this->Category_model->check_product($id);
+//        var_dump($check_product);die;
         if ($row) {
-          if ($check_sub_cat<1) {
+          if ($check_sub_cat<1 && $check_product<1) {
             $this->Category_model->delete($id);
             $this->session->set_flashdata('message', $row->name. ' Berhasil Dihapus');
             redirect(site_url('admin/category'));
           }elseif($check_sub_cat>=1) {
-            $this->session->set_flashdata('message', 'Gagal Menghapus Karena Kategori Ini Masih Terikat Dengan Sub Kategori');
+            $this->session->set_flashdata('message', 'Gagal !! Masih Terikat Dengan Sub Kategori');
+            redirect(site_url('admin/category'));
+          }elseif($check_product>=1) {
+            $this->session->set_flashdata('message', 'Gagal !! Masih Ada Produk Pada Kategori Ini');
             redirect(site_url('admin/category'));
           }
         } else {
